[ROUTING] Move document related dynamic route matching to a dedicated document handler #1374
Keeping all document related code in DynmicRouteProvider seems bloated
and makes the dynamic provider unreadable. This refactors document
routing into a DocumentRouteHandler which acts upon a
DynamicRouteContext (Request, path, originalPath) to find a matching
document route.

Still WIP as handler is not wired to the dynamic route provider yet.
